+++ packages/sanes-chrome-extension/src/extension/background/externalHandler/requestHandler.ts
