+++ packages/sanes-chrome-extension/src/extension/background/externalHandler/senderWhitelist.ts
